// libs/services/src/lib/models/enums/Page.ts
export enum Page {
    None = 'NONE',
    Home = 'home',
    Team = 'team',
    Updates = 'updates',
    Resources = 'resources',
    University = 'university',
    Catalog = 'catalog',
    CatalogItemDetail = 'catalog-item-detail',
    CatalogItemViewer = 'catalog-item-viewer',
    PageNotFound = 'pagenotfound'
}
export enum ChildPage {
    None = 'NONE',
    Home = 'home',
    About = 'about',
    Catalog = 'catalog',
    CatalogItemDetail = 'catalog-item-detail',
    CatalogItemViewer = 'catalog-item-viewer',
    PageNotFound = 'pagenotfound'
}