// libs/services/src/lib/models/enums/PageInkPosition.ts
export enum PageInkPosition {
    None = -100,
    Home = 34,
    Team = 120,
    Updates = 238,
    Resources = 413,
    University = -155,
    PageNotFound = -155
}
export enum PageInkWidth {
    None = 0,
    Home = 52,
    Team = 82,
    Updates = 140,
    Resources = 90,
    University = 6,
    PageNotFound = 0
}